+++ java-opentimestamps/src/main/java/com/eternitywall/ots/crypto/Pack.java
@@ -65,7 +65,7 @@ public abstract class Pack
     {
         int hi = bigEndianToInt(bs, off);
         int lo = bigEndianToInt(bs, off + 4);
-        return ((long)(hi & 0xffffffffL) << 32) | (long)(lo & 0xffffffffL);
+        return ((hi & 0xffffffffL) << 32) | lo & 0xffffffffL;
     }
 
     public static void bigEndianToLong(byte[] bs, int off, long[] ns)
@@ -199,7 +199,7 @@ public abstract class Pack
     {
         int lo = littleEndianToInt(bs, off);
         int hi = littleEndianToInt(bs, off + 4);
-        return ((long)(hi & 0xffffffffL) << 32) | (long)(lo & 0xffffffffL);
+        return ((hi & 0xffffffffL) << 32) | lo & 0xffffffffL;
     }
 
     public static void littleEndianToLong(byte[] bs, int off, long[] ns)
